+++ lib/src/cache/cache_view.dart
@@ -1,15 +1,15 @@
 import 'package:restrr/restrr.dart';
 
 class RestrrEntityCacheView<T extends RestrrEntity> {
-  final Map<ID, T> _cache = {};
+  final Map<Id, T> _cache = {};
 
-  T? get(ID id) => _cache[id];
+  T? get(Id id) => _cache[id];
 
   T cache(T value) => _cache[value.id] = value;
 
-  T? remove(ID id) => _cache.remove(id);
+  T? remove(Id id) => _cache.remove(id);
 
   void clear() => _cache.clear();
 
-  bool contains(ID id) => _cache.containsKey(id);
+  bool contains(Id id) => _cache.containsKey(id);
 }