+++ lib/src/entities/restrr_entity.dart
@@ -1,6 +1,6 @@
 import '../../restrr.dart';
 
-typedef ID = int;
+typedef Id = int;
 
 /// The base class for all Restrr entities.
 /// This simply provides a reference to the Restrr instance.
@@ -8,14 +8,14 @@ abstract class RestrrEntity {
   /// A reference to the Restrr instance.
   Restrr get api;
 
-  ID get id;
+  Id get id;
 }
 
 class RestrrEntityImpl implements RestrrEntity {
   @override
   final Restrr api;
   @override
-  final ID id;
+  final Id id;
 
   const RestrrEntityImpl({
     required this.api,