+++ lib/src/service/session_service.dart
@@ -5,25 +5,34 @@ import '../../restrr.dart';
 class SessionService extends ApiService {
   const SessionService({required super.api});
 
-  static Future<RestResponse<Session>> create(Uri uri, String username, String password, {String? sessionName}) async {
-    return RequestHandler.request(
+  Future<RestResponse<Session>> create(String username, String password, {String? sessionName}) async {
+    return request(
         route: SessionRoutes.create.compile(),
         body: {
           'username': username,
           'password': password,
           'session_name': sessionName,
         },
-        mapper: (json) => EntityBuilder.buildSession(json),
+        mapper: (json) => api.entityBuilder.buildSession(json),
         errorMap: {
           404: RestrrError.invalidCredentials,
           401: RestrrError.invalidCredentials,
         });
   }
 
-  Future<RestResponse<bool>> logout() async {
+  Future<RestResponse<Session>> refresh() async {
+    return request(
+        route: SessionRoutes.create.compile(),
+        mapper: (json) => api.entityBuilder.buildSession(json),
+        errorMap: {
+          404: RestrrError.invalidCredentials,
+          401: RestrrError.invalidCredentials,
+        });
+  }
+
+  Future<RestResponse<bool>> delete() async {
     return noResponseRequest(route: SessionRoutes.deleteCurrent.compile(), errorMap: {
       401: RestrrError.notSignedIn,
     });
   }
-
 }